# PERF: Index.get_loc